finished client profile requesst

+++ front-end/src/app/services/auth/local-storage-manager.service.ts
@@ -7,36 +7,49 @@ export class LocalStorageManagerService {
 
   constructor() { }
 
+  private isLocalStorageAvailable(): boolean {
+    return typeof window !== 'undefined' && !!window.localStorage;
+  }
+
   setItem(key: string, value: string): void {
-    try {
-      localStorage.setItem(key, value);
-    } catch (e) {
-      console.error(`Error saving to localStorage: ${e}`);
+    this.clearStorage();
+    if (this.isLocalStorageAvailable()) {
+      try {
+        localStorage.setItem(key, value);
+      } catch (e) {
+        console.error(`Error saving to localStorage: ${e}`);
+      }
     }
   }
 
   getItem(key: string): string | null {
-    try {
-      return localStorage.getItem(key);
-    } catch (e) {
-      console.error(`Error reading from localStorage: ${e}`);
-      return null;
+    if (this.isLocalStorageAvailable()) {
+      try {
+        return localStorage.getItem(key);
+      } catch (e) {
+        console.error(`Error reading from localStorage: ${e}`);
+      }
     }
+    return null;
   }
 
   removeItem(key: string): void {
-    try {
-      localStorage.removeItem(key);
-    } catch (e) {
-      console.error(`Error removing from localStorage: ${e}`);
+    if (this.isLocalStorageAvailable()) {
+      try {
+        localStorage.removeItem(key);
+      } catch (e) {
+        console.error(`Error removing from localStorage: ${e}`);
+      }
     }
   }
 
   clearStorage(): void {
-    try {
-      localStorage.clear();
-    } catch (e) {
-      console.error(`Error clearing localStorage: ${e}`);
+    if (this.isLocalStorageAvailable()) {
+      try {
+        localStorage.clear();
+      } catch (e) {
+        console.error(`Error clearing localStorage: ${e}`);
+      }
     }
   }
